fix(agent): avoid eager fork capability probes
